dev 0.6.0j
Functional Battle System

Notable bugs
 + Player attack spam
 + Enemy animation attack spam

+++ Siegeldun Game/Assets/Scripts/Enemy.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
 
     void Start()
     {
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
